refactor: add time records repository layer and move code from service there
