+++ plugins/services.ts
@@ -11,15 +11,18 @@ import ContractsService from '~/services/contracts-service';
 import BridgeService from '~/services/bridge-service';
 import TimesheetsRepository from '~/api/timesheets/timesheets-repository';
 import WorkSchemeRepository from '~/api/work-scheme/work-scheme-repository';
+import TimeRecordsRepository from '~/api/records/time-records-repository';
 
 export default defineNuxtPlugin(({$fire, $fireModule, $axios}, inject) => {
+  const travelRecordsService = new TravelRecordsService($fire);
   const timesheetsRepository = new TimesheetsRepository($fire);
   const workSchemeRepository = new WorkSchemeRepository($axios);
+  const timeRecordsRepository = new TimeRecordsRepository($fire, $axios);
 
   inject('contractsService', new ContractsService($axios));
   inject('customersService', new CustomersService($fire, $fireModule));
-  inject('timeRecordsService', new TimeRecordsService($fire, $axios));
-  inject('travelRecordsService', new TravelRecordsService($fire));
+  inject('timeRecordsService', new TimeRecordsService(timeRecordsRepository));
+  inject('travelRecordsService', travelRecordsService);
   inject('employeesService', new EmployeesService($fire, $fireModule));
   inject('teamsService', new TeamsService($fire, $fireModule));
   inject('workSchemeService', new WorkSchemeService(workSchemeRepository));
